[CONFIG] [Github Actions] build / test environments splitted by OS.

+++ src/lib/exercises/include/exercises/hackerrank/warmup/a_very_big_sum.hpp
@@ -3,5 +3,5 @@
 #pragma once
 
 namespace hackerrank::warmup {
-long aVeryBigSum(const std::vector<long> &ar);
+long long aVeryBigSum(const std::vector<long> &ar);
 }
